--- a/tests/appsec/iast/taint_tracking/test_native_taint_range.py
+++ b/tests/appsec/iast/taint_tracking/test_native_taint_range.py
@@ -542,12 +542,12 @@ def reset_contexts_loop():
 
 
 async def async_reset_context_loop(task_id: int):
-    await asyncio.sleep(0.01)
+    await asyncio.sleep(0.03)
     return reset_context_loop()
 
 
 async def async_reset_contexts_loop(task_id: int):
-    await asyncio.sleep(0.01)
+    await asyncio.sleep(0.02)
     return reset_contexts_loop()
 
 
@@ -556,7 +556,7 @@ def test_race_conditions_threads(caplog, telemetry_writer):
     destroying contexts
     """
     pool = ThreadPool(processes=3)
-    results_async = [pool.apply_async(reset_context_loop) for _ in range(70)]
+    results_async = [pool.apply_async(reset_context_loop) for _ in range(20)]
     _ = [res.get() for res in results_async]
 
     log_messages = [record.message for record in caplog.get_records("call")]
@@ -577,7 +577,7 @@ def test_race_conditions_reset_contexts_threads(caplog, telemetry_writer):
         dict(_iast_debug=True)
     ), caplog.at_level(logging.DEBUG):
         pool = ThreadPool(processes=3)
-        results_async = [pool.apply_async(reset_contexts_loop) for _ in range(70)]
+        results_async = [pool.apply_async(reset_contexts_loop) for _ in range(20)]
         _ = [res.get() for res in results_async]
 
         log_messages = [record.message for record in caplog.get_records("call")]
@@ -592,7 +592,7 @@ async def test_race_conditions_reset_contex_async(caplog, telemetry_writer):
     """we want to validate context is working correctly among multiple request and no race condition creating and
     destroying contexts
     """
-    tasks = [async_reset_context_loop(i) for i in range(50)]
+    tasks = [async_reset_context_loop(i) for i in range(7)]
 
     results = await asyncio.gather(*tasks)
     assert results
@@ -611,7 +611,7 @@ async def test_race_conditions_reset_contexs_async(caplog, telemetry_writer):
     """we want to validate context is working correctly among multiple request and no race condition creating and
     destroying contexts
     """
-    tasks = [async_reset_contexts_loop(i) for i in range(20)]
+    tasks = [async_reset_contexts_loop(i) for i in range(10)]
 
     results = await asyncio.gather(*tasks)
     assert results